[common] enabled `verbatimModuleSyntax`
