+++ packages/common/src/localization/LocalizedValidationErrors.ts
@@ -1,4 +1,4 @@
-import {
+import type {
     IErrorsLocalization,
     ValidationErrorsFactory,
     ValidationErrorsStrings,
