+++ packages/common/src/types/deep.ts
@@ -1,4 +1,4 @@
-import { AnyFunction } from './misc.js';
+import type { AnyFunction } from './misc.js';
 
 export type DeepReadonly<T> = T extends Array<infer U>
     ? ReadonlyArray<DeepReadonly<U>>
