+++ packages/common/src/types/ident.ts
@@ -4,5 +4,3 @@ export type Ident<T, K extends keyof any = string> = T & {
 };
 
 export type IdentAny = Ident<object>;
-
-export default Ident;
