+++ packages/common/src/validation/types.ts
@@ -1,4 +1,4 @@
-import { ValidationErrors } from './ValidationErrors.js';
+import { type ValidationErrors } from './ValidationErrors.js';
 
 export type ValidatorFunction<T = string, TErrors = ValidationErrors, TContext = any> = (val: T, ctx?: TContext) => TErrors;
 export type ValidatorFunctionAsync<T = string, TErrors = ValidationErrors, TContext = any> = (val: T, ctx?: TContext) => Promise<TErrors>;
@@ -7,15 +7,6 @@ export type ValidationConfig<T, TErrors> = { [P in keyof T]?: ValidatorFunction<
 export type ValidationResults<T, TErrors> = { [P in keyof T]?: TErrors };
 
 export type WrapperFunction = (val: ValidatorFunction) => ValidatorFunction;
-
-export class ValidationError<TErrors = ValidationErrors> extends Error {
-    readonly code: TErrors;
-
-    constructor(message: string, code: TErrors) {
-        super(message);
-        this.code = code;
-    }
-}
 
 export interface ValidationVoid {
     readonly error?: string | null | undefined;
