+++ tsconfig.json
@@ -8,6 +8,7 @@
         "skipLibCheck": true,
         "sourceMap": true,
         "isolatedModules": true,
+        "verbatimModuleSyntax": true,
         "esModuleInterop": true,
         "target": "ES2022",
         "lib": ["ESNext"],
